% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_latex_modifiers.R
\name{latex_scale}
\alias{latex_scale}
\title{Scale full table or individual columns for latex table output}
\usage{
latex_scale(latex_code, scale_factor, align = c("l", "r", "c"))
}
\arguments{
\item{latex_code}{knit_asis chr string generated by \code{as_latex()} on a gt table object.}

\item{scale_factor}{proportion to shrink or enlarge. can be a numeric value to scale the full table or a list of vectors in the format \verb{list(c(column number, scale factor))} to apply scaling to specific column(s).}

\item{align}{align direction to align the full table on the page.}
}
\description{
Scale full table or individual columns for latex table output
}
\examples{
df <- dplyr::tribble(
 ~ fruit, ~ count,  ~ color,
 'apple',       1,    'red',
'banana',       2,  'yellow'
)

# shrink column 1 by 0.5 and column 3 by 0.7.
df \%>\%
gt() \%>\%
as_latex() \%>\%
latex_scale(scale_factor = list(c(1, 0.5),
                                c(3, 0.7)))

# shrink all the columns by 0.5.
df \%>\%
gt() \%>\%
as_latex() \%>\%
latex_scale(scale_factor = 0.5)

}
