% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_latex_modifiers.R
\name{lscape}
\alias{lscape}
\title{Orient a LateX table in landscape}
\usage{
lscape(latex_code)
}
\arguments{
\item{latex_code}{knit_asis chr string generated by \code{as_latex()} on a gt table object or a list of knit_asis chr strings.}
}
\description{
Orient a LateX table in landscape
}
\examples{
df <- dplyr::tribble(
 ~ fruit, ~ count,  ~ color,
 'apple',       1,    'red',
'banana',       2,  'yellow'
)

# single table on a single landscape page

df \%>\%
gt() \%>\%
as_latex() \%>\%
lscape()

# multiple tables on a single landscape page
tbl <- df \%>\%
       gt() \%>\%
       as_latex()

(tbl + tbl + tbl) \%>\% lscape()

 #      -- or --

list(tbl, tbl, tbl) \%>\% lscape()

}
