% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaa.R
\name{mrggt_options}
\alias{mrggt_options}
\title{Global options to set for mrggt that effect the latex rendering}
\usage{
mrggt_options(...)
}
\arguments{
\item{...}{args passed on to assign function. possible values to assign are:
\itemize{
\item \strong{pagemargin}: numeric vector in inches with format \verb{c(left margin, right margin)}; default is \code{c(1, 1)}
\item \strong{papersize}: character; default \code{'letter'}; options:
\itemize{
\item \emph{half letter}: 5.5 x 8.0 in
\item \emph{letter}: 8.5 x 11.0 in
\item \emph{legal}: 8.5 x 14.0 in
\item \emph{junior legal}: 5.0 x 8.0 in
\item \emph{ledger}: 11.0 x 17.0 in
}
}}
}
\description{
Global options to set for mrggt that effect the latex rendering
}
\examples{
# set left & right margins to 3in & 4in
# no line breaks
# change paper to legal

mrggt_options('pagemargin' = c(3, 4),
             'papersize' = 'legal')

}
