% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaa.R
\name{mrggtOptions}
\alias{mrggtOptions}
\title{Global options to set for mrggt that effect the latex rendering}
\usage{
mrggtOptions(...)
}
\arguments{
\item{...}{args passed on to assign function. possible values to assign are:
\itemize{
\item \strong{margin}: numeric vector with values in inches in format \verb{c(left margin, right margin)}; default is \code{c(1, 1)}
\item \strong{column.sep}: numeric value in pt; default is 3pt.
\item \strong{line.breaks}: logical; allow line breaks in table; default is \code{TRUE} (recommended)
}}
}
\description{
Global options to set for mrggt that effect the latex rendering
}
\examples{
# set left & right margins to 1in & 2in
# no line breaks
# change column separation in table to 2pt.

mrggtOptions('margin' = c(3, 4),
             'line.breaks' = FALSE,
             'column.sep' = 2)

}
