% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_latex_modifiers.R
\name{preamble}
\alias{preamble}
\title{Extract a \strong{gt knit_asis} object's preamble or output a \strong{gt knit_asis} object with preamble included.}
\usage{
preamble(tex, full_tex = TRUE)
}
\arguments{
\item{tex}{a knit_asis object that is created using the \code{as_latex()} function.}

\item{full_tex}{output only the preamble or the preamble with tex code (including document declaration); default is TRUE - output tex code + preamble}
}
\description{
Extract a \strong{gt knit_asis} object's preamble or output a \strong{gt knit_asis} object with preamble included.
}
\examples{
# Use `gtcars` to create a gt table;
# add a header and then export just preamble
# as tex file

  gtcars \%>\%
  dplyr::select(mfr, model) \%>\%
  dplyr::slice(1:2) \%>\%
  gt() \%>\%
  tab_header(
    title = md("Data listing from **gtcars**"),
    subtitle = md("`gtcars` is an R dataset")
  ) \%>\%
  as_latex() \%>\%
  preamble(full_tex = FALSE)

}
