% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_save_latex.R
\name{save_latex}
\alias{save_latex}
\title{Output a \strong{gt knit_asis} or \strong{gt preamble} object to a tex file}
\usage{
## Default S3 method:
 save_latex(tex, path, ...)

 ## S3 method for class `knit_asis`:
 save_latex(tex, path, preamble = TRUE, append = FALSE)

 ## S3 method for class `preamble`:
 save_latex(tex, path, append = FALSE)
}
\arguments{
\item{tex}{a \code{knit_asis} object that is created using the \code{as_latex()} function or a \code{preamble} object created using \code{preamble()} function.}

\item{path}{path or connection to write to.}

\item{...}{further arguments passed to or used by methods.}

\item{preamble}{include full preamble with document declaration and document environment wrapping table; default is \code{FALSE}. Ignored if \code{tex} is \code{preamble} object.}

\item{append}{If \code{FALSE}, will overwrite existing file. If \code{TRUE}, will append to existing file. In both cases, if file does not exist a new file is created.}
}
\description{
Output a \strong{gt knit_asis} or \strong{gt preamble} object to a tex file
}
\examples{
# Use `gtcars` to create a gt table;
# add a header and then export as
# tex file with preamble and document declarations

  gtcars \%>\%
  dplyr::select(mfr, model) \%>\%
  dplyr::slice(1:2) \%>\%
  gt() \%>\%
  tab_header(
    title = md("Data listing from **gtcars**"),
    subtitle = md("`gtcars` is an R dataset")
  ) \%>\%
  as_latex() \%>\%
  save_latex('tex.tex', preamble = TRUE)

# Use `gtcars` to create a gt table;
# add a header and then export
# just the preamble (package declarations + option settings) to file.

  gtcars \%>\%
  dplyr::select(mfr, model) \%>\%
  dplyr::slice(1:2) \%>\%
  gt() \%>\%
  tab_header(
    title = md("Data listing from **gtcars**"),
    subtitle = md("`gtcars` is an R dataset")
  ) \%>\%
  as_latex() \%>\%
  preamble(full_tex = FALSE) \%>\%
  save_latex('preamble.tex')

}
