% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_create_modify.R
\name{tab_caption}
\alias{tab_caption}
\title{Add a caption to a LaTeX table}
\usage{
tab_caption(data, caption)
}
\arguments{
\item{data}{A table object that is created using the \code{\link[=gt]{gt()}} function.}

\item{caption}{the caption to be set within \verb{\\caption\{\}} in the \code{longtable} environment. \code{label} can also be passed.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
Set a caption within the \code{longtable} LaTeX environment.
}
\examples{
# Use `gtcars` to create a gt table;
# Group several columns related to car
# add a caption of 'Gtcars table'
# with the label `performance`
tab_1 <-
  gtcars \%>\%
  dplyr::select(
    -mfr, -trim, bdy_style, drivetrain,
    -drivetrain, -trsmn, -ctry_origin
  ) \%>\%
  dplyr::slice(1:8) \%>\%
  gt(rowname_col = "model") \%>\%
  tab_caption(caption = "Gtcars table\\\\label{performance}")

}
